refactor: combine player and player controller into player manager
