+++ Assets/Prefabs/Player.prefab
@@ -48,7 +48,6 @@ GameObject:
   - component: {fileID: 5510391779424095209}
   - component: {fileID: 5281355154443792004}
   - component: {fileID: 8322529334511602896}
-  - component: {fileID: 2972697643043051729}
   - component: {fileID: 3100164559581419505}
   m_Layer: 3
   m_Name: Player
@@ -255,6 +254,7 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: e1d434b47837d804eb8efff4c931150f, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  canMove: 1
   jumpCount: 0
 --- !u!114 &5510391779424095209
 MonoBehaviour:
@@ -300,20 +300,6 @@ MonoBehaviour:
   MaxHealth: 125
   currentHealth: 0
   Healthbar: {fileID: 7748277353179727481}
---- !u!114 &2972697643043051729
-MonoBehaviour:
-  m_ObjectHideFlags: 0
-  m_CorrespondingSourceObject: {fileID: 0}
-  m_PrefabInstance: {fileID: 0}
-  m_PrefabAsset: {fileID: 0}
-  m_GameObject: {fileID: 5997975773698609013}
-  m_Enabled: 1
-  m_EditorHideFlags: 0
-  m_Script: {fileID: 11500000, guid: a3bacd01c76de8a4ea41351eea03f50b, type: 3}
-  m_Name: 
-  m_EditorClassIdentifier: 
-  deathYThreshold: -20
-  spawnPoint: {fileID: 0}
 --- !u!114 &3100164559581419505
 MonoBehaviour:
   m_ObjectHideFlags: 0
@@ -326,6 +312,8 @@ MonoBehaviour:
   m_Script: {fileID: 11500000, guid: 2cbc3acde7a6a534cb013f828c39ac13, type: 3}
   m_Name: 
   m_EditorClassIdentifier: 
+  deathYThreshold: -20
+  spawnPoint: {fileID: 0}
 --- !u!1 &6389031454060092190
 GameObject:
   m_ObjectHideFlags: 0
